#35: Maintain support for swift <5.9
