+++ Package.swift
@@ -1,48 +1,7 @@
-// swift-tools-version:5.9
+// swift-tools-version:5.7
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
-import CompilerPluginSupport
-
-
-var targets: [Target] = [
-    .target(
-        name: "URLMock",
-        dependencies: ["Relax"]
-    ),
-    .testTarget(
-        name: "RelaxTests",
-        dependencies: ["Relax", "URLMock"]
-    ),
-]
-
-var dependencies = [Package.Dependency]()
-
-// Macros do not currently compile on windows when building tests: https://github.com/apple/swift-package-manager/issues/7174
-#if canImport(XCTest) && os(Windows)
-targets.append(.target(name: "Relax"))
-#else
-targets.append(
-    contentsOf: [
-        .target(name: "Relax", dependencies: ["RelaxMacros"]),
-        .macro(
-            name: "RelaxMacros",
-            dependencies: [
-                .product(name: "SwiftSyntaxMacros", package: "swift-syntax"),
-                .product(name: "SwiftCompilerPlugin", package: "swift-syntax")
-            ]
-        ),
-        .testTarget(
-            name: "RelaxMacrosTests",
-            dependencies: ["RelaxMacros", .product(name: "SwiftSyntaxMacrosTestSupport", package: "swift-syntax")]
-        )
-    ]
-)
-dependencies = [
-    // Depend on the Swift 5.9 release of SwiftSyntax
-    .package(url: "https://github.com/apple/swift-syntax.git", from: "509.0.0"),
-]
-#endif
 
 let package = Package(
     name: "Relax",
@@ -61,6 +20,22 @@ let package = Package(
             name: "URLMock",
             targets: ["URLMock"])
     ],
-    dependencies: dependencies,
-    targets: targets
+    dependencies: [
+        // Dependencies declare other packages that this package depends on.
+        // .package(url: /* package url */, from: "1.0.0"),
+    ],
+    targets: [
+        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
+        // Targets can depend on other targets in this package, and on products in packages which this package depends on.
+        .target(
+            name: "Relax",
+            dependencies: []),
+        .target(
+            name: "URLMock",
+            dependencies: ["Relax"]
+        ),
+        .testTarget(
+            name: "RelaxTests",
+            dependencies: ["Relax", "URLMock"]),
+    ]
 )
